requirements.txt
Django==4.0.4
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
whitenoise==6.0.0
psycopg2-binary==2.9.3
django-tables2==2.4.1
django-filter==21.1
django-phonenumbers==1.0.1
django-phonenumber-field==6.1.0